Small refactoring & fixups, cmake cleanup
