+++ CMakeLists.txt
@@ -1,18 +1,47 @@
-set(SOURCES "src/Event/EventScanner.cpp" "src/Object/Class.cpp" "src/Devices/AW9523.cpp" "src/Drivers/Input/InputAW.cpp" "src/Drivers/Output/OutputDigAW.cpp" "src/Drivers/Output/OutputCurrAW.cpp" "src/Drivers/Input/InputGPIO.cpp" "src/Drivers/Input/InputTouchGPIO.cpp" "src/Drivers/Output/OutputGPIO.cpp" "src/Services/ButtonInput.cpp" "src/Periphery/GPIOPeriph.cpp" "src/Periphery/I2C.cpp" "src/Util/StateMachine/State.cpp" "src/Util/StateMachine/StateMachine.cpp" "src/Containers/Archive.cpp" "src/Misc/Singleton.cpp" "src/Memory/ObjectMemory.cpp" "src/Object/Object.cpp" "src/Core/Application.cpp" "src/Memory/ObjectManager.cpp" "src/Memory/GarbageCollector.cpp" "src/Entity/Entity.cpp" "src/Entity/AsyncEntity.cpp" "src/Entity/SyncEntity.cpp" "src/Statics/ApplicationStatics.cpp" "src/Thread/Threaded.cpp" "src/Util/stdafx.cpp" CACHE INTERNAL "" FORCE)
+set(SRC_DIRS
+        "${CMAKE_CURRENT_SOURCE_DIR}/src/Containers"
+        "${CMAKE_CURRENT_SOURCE_DIR}/src/Core"
+        "${CMAKE_CURRENT_SOURCE_DIR}/src/Devices"
+        "${CMAKE_CURRENT_SOURCE_DIR}/src/Drivers"
+        "${CMAKE_CURRENT_SOURCE_DIR}/src/Entity"
+        "${CMAKE_CURRENT_SOURCE_DIR}/src/Event"
+        "${CMAKE_CURRENT_SOURCE_DIR}/src/Log"
+        "${CMAKE_CURRENT_SOURCE_DIR}/src/Memory"
+        "${CMAKE_CURRENT_SOURCE_DIR}/src/Misc"
+        "${CMAKE_CURRENT_SOURCE_DIR}/src/Object"
+        "${CMAKE_CURRENT_SOURCE_DIR}/src/Periphery"
+        "${CMAKE_CURRENT_SOURCE_DIR}/src/Services"
+        "${CMAKE_CURRENT_SOURCE_DIR}/src/Statics"
+        "${CMAKE_CURRENT_SOURCE_DIR}/src/Thread"
+        "${CMAKE_CURRENT_SOURCE_DIR}/src/Util"
+        "${CMAKE_CURRENT_SOURCE_DIR}/src/FileSystem"
+        CACHE INTERNAL "")
 
-file(GLOB_RECURSE FS_SOURCES "src/FS/**.cpp" "src/FS/**.c")
+set(EXCLUDED_ITEMS
+        "${CMAKE_CURRENT_SOURCE_DIR}/src/Devices/Display.h"
+        "${CMAKE_CURRENT_SOURCE_DIR}/src/Devices/Display.cpp"
+        CACHE INTERNAL "")
+
+foreach(SRC IN LISTS SRC_DIRS)
+    file(GLOB_RECURSE DIR_SOURCES "${SRC}/**.h" "${SRC}/**.cpp" "${SRC}/**.hpp" "${SRC}/**.inl")
+    list(APPEND SOURCES ${DIR_SOURCES})
+endforeach()
+
+foreach(SRC IN LISTS EXCLUDED_ITEMS)
+    list(REMOVE_ITEM SOURCES "${SRC}")
+endforeach()
 
 if (CONFIG_CMF_GRAPHICS)
-    file(GLOB_RECURSE GRAPHICS_SOURCES "src/LV_Interface/**.cpp" "src/LV_Interface/**.c")
-    list(APPEND GRAPHICS_SOURCES "src/Devices/Display.cpp")
+    file(GLOB_RECURSE GRAPHICS_SOURCES "${CMAKE_CURRENT_SOURCE_DIR}/src/LV_Interface/**.cpp" "${CMAKE_CURRENT_SOURCE_DIR}/src/LV_Interface/**.c" "${CMAKE_CURRENT_SOURCE_DIR}/src/LV_Interface/**.hpp" "${CMAKE_CURRENT_SOURCE_DIR}/src/LV_Interface/**.h")
+    list(APPEND GRAPHICS_SOURCES "${CMAKE_CURRENT_SOURCE_DIR}/src/Devices/Display.cpp" "${CMAKE_CURRENT_SOURCE_DIR}/src/Devices/Display.h")
 
     list(APPEND SOURCES ${GRAPHICS_SOURCES})
 endif()
 
-list(APPEND SOURCES ${FS_SOURCES})
-
-
 file(GLOB_RECURSE LIBS "lib/*/src/**.cpp" "lib/*/src/**.c")
-set(LIBS_INCL "lib/glm/glm")
 
-idf_component_register(SRCS ${SOURCES}  ${LIBS} INCLUDE_DIRS "src"  ${LIBS_INCL} REQUIRES esp_timer driver spiffs)
+set(LIBS_INCL "lib/glm/glm" CACHE INTERNAL "")
+
+set(REQUIRED_COMPONENTS "esp_timer" "driver" "spiffs" CACHE INTERNAL "")
+
+idf_component_register(SRCS ${SOURCES}  ${LIBS} INCLUDE_DIRS "src"  ${LIBS_INCL} REQUIRES ${REQUIRED_COMPONENTS})
