+++ src/Devices/Display.h
@@ -1,8 +1,9 @@
-#ifndef CLOCKSTAR_FIRMWARE_DISPLAY_H
-#define CLOCKSTAR_FIRMWARE_DISPLAY_H
+#ifndef CMF_DISPLAY_H
+#define CMF_DISPLAY_H
 
 #include "Object/Object.h"
 #include <LovyanGFX.h>
+
 typedef lgfx::LGFX_Sprite Sprite;
 
 /**
@@ -10,6 +11,7 @@ typedef lgfx::LGFX_Sprite Sprite;
  */
 class Display : public Object {
 	GENERATED_BODY(Display, Object)
+
 public:
 	Display() = default;
 	Display(lgfx::Bus_SPI::config_t busConfig, lgfx::Panel_Device::config_t panelConfig, std::function<void(Sprite&)> canvasInit = {});
@@ -31,5 +33,4 @@ private:
 	Sprite canvas;
 };
 
-
-#endif //CLOCKSTAR_FIRMWARE_DISPLAY_H
+#endif //CMF_DISPLAY_H
