+++ src/FileSystem/FileCache.h
@@ -1,8 +1,6 @@
 #ifndef CMF_FILECACHE_H
 #define CMF_FILECACHE_H
 
-
-#include <unordered_map>
 #include "File.h"
 
 class FileCache {
@@ -12,8 +10,6 @@ public:
 
 	virtual void load() = 0;
 	virtual void unload() = 0;
-
 };
-
 
 #endif //CMF_FILECACHE_H
